Do not show "you are not logged in" message if account just hasn't loaded. Fix #158.
